// sanity.types.ts
export type Product = {
  _id: string;
  _type: "product";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  images?: Array<{
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
    _key: string;
  }>;
  description?: string;
  price?: number;
  discount?: number;
  categories?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "category";
  }>;
  stock?: number;
  brand?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "brand";
  };
  status?: "new" | "hot" | "sale" | "featured" | "limited";
  productType?:
    | "pen"
    | "pencil"
    | "marker"
    | "highlighter"
    | "notebook"
    | "diary"
    | "planner"
    | "sticky_notes"
    | "eraser"
    | "ruler"
    | "stapler"
    | "paper_clips"
    | "binder"
    | "file_folder"
    | "calculator"
    | "scissors"
    | "glue"
    | "tape"
    | "ball"
    | "bat"
    | "racket"
    | "sports_shoes"
    | "jersey"
    | "shorts"
    | "gloves"
    | "helmet"
    | "knee_pads"
    | "water_bottle"
    | "gym_bag"
    | "whistle"
    | "stopwatch"
    | "jump_rope"
    | "yoga_mat"
    | "gift_card"
    | "gift_box"
    | "greeting_card"
    | "photo_frame"
    | "keychain"
    | "mug"
    | "t_shirt"
    | "cushion"
    | "wall_art"
    | "candle"
    | "jewelry"
    | "watch"
    | "toy"
    | "puzzle"
    | "board_game"
    | "face_cream"
    | "face_wash"
    | "moisturizer"
    | "sunscreen"
    | "serum"
    | "toner"
    | "cleanser"
    | "face_mask"
    | "foundation"
    | "concealer"
    | "lipstick"
    | "lip_balm"
    | "mascara"
    | "eyeliner"
    | "eyeshadow"
    | "blush"
    | "nail_polish"
    | "perfume"
    | "body_lotion"
    | "shampoo"
    | "conditioner"
    | "hair_oil"
    | "hair_mask"
    | "others";
  subcategory?: string;
  isFeatured?: boolean;
  weight?: number;
  dimensions?: {
    length?: number;
    width?: number;
    height?: number;
  };
  tags?: Array<string>;
};
// ...
export type Category = {
  _id: string;
  _type: "category";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  slug?: Slug;
  description?: string;
  range?: number;
  featured?: boolean;
  productCount?: number;
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
};
// ...
export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// Source: ./sanity/queries/query.ts
// Variable: BRANDS_QUERY
// Query: *[_type=='brand'] | order(name asc)
export type BRANDS_QUERYResult = Array<{
  _id: string;
  _type: "brand";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  slug?: Slug;
  description?: string;
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  categories?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "category";
  }>;
}>;